+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/SequencingContainerModelRestController.java
@@ -2,18 +2,24 @@ package uk.ac.bbsrc.tgac.miso.webapp.controller.rest;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import javax.ws.rs.core.Response.Status;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
 import uk.ac.bbsrc.tgac.miso.core.service.SequencingContainerModelService;
 import uk.ac.bbsrc.tgac.miso.dto.Dtos;
 import uk.ac.bbsrc.tgac.miso.dto.SequencingContainerModelDto;
@@ -51,6 +57,23 @@ public class SequencingContainerModelRestController extends RestController {
   @ResponseStatus(HttpStatus.NO_CONTENT)
   public void bulkDelete(@RequestBody(required = true) List<Long> ids) throws IOException {
     RestUtils.bulkDelete("Container Model", ids, containerModelService);
+  }
+
+  @GetMapping("/search")
+  public @ResponseBody List<SequencingContainerModelDto> search(@RequestParam("platformType") String platformType,
+      @RequestParam("q") String query) throws IOException {
+    PlatformType platform = validatePlatformType(platformType);
+    return containerModelService.find(platform, query).stream()
+        .map(Dtos::asDto)
+        .collect(Collectors.toList());
+  }
+
+  private PlatformType validatePlatformType(String platformType) {
+    try {
+      return PlatformType.valueOf(platformType);
+    } catch (IllegalArgumentException e) {
+      throw new RestException("Invalid platform type: " + platformType, Status.BAD_REQUEST);
+    }
   }
 
 }