+++ sqlstore/src/test/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateSequencingContainerModelDaoTest.java
@@ -144,4 +144,11 @@ public class HibernateSequencingContainerModelDaoTest extends AbstractDAOTest {
     assertEquals(4, dao.getUsage(model));
   }
 
+  @Test
+  public void testGetUsagePlatform() throws Exception {
+    SequencingContainerModel containerModel = (SequencingContainerModel) currentSession().get(SequencingContainerModel.class, 1L);
+    InstrumentModel instrumentModel = (InstrumentModel) currentSession().get(InstrumentModel.class, 16L);
+    assertEquals(4, dao.getUsage(containerModel, instrumentModel));
+  }
+
 }